Backend

+++ project-manager/src/main/java/ateam/CompletionlevelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ateam.Completionlevel;
 
-public interface CompletionlevelRepository extends CrudRepository<Completionlevel, Long>{
+public interface CompletionlevelRepository extends CrudRepository<Completionlevel, Integer>{
 
 }
