+++ project-manager/src/main/java/ateam/ProjectRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ateam.Project;
 
-public interface ProjectRepository extends CrudRepository<Project, Long>{
+public interface ProjectRepository extends CrudRepository<Project, Integer>{
 
 }
