+++ project-manager/src/main/java/ateam/TaskRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import ateam.Task;
 
-public interface TaskRepository extends CrudRepository<Task, Long>{
+public interface TaskRepository extends CrudRepository<Task, Integer>{
 
 }
